+++ components/klab.component.ml/src/main/java/org/integratedmodelling/ml/adapters/WekaImporter.java
@@ -1,8 +1,10 @@
 package org.integratedmodelling.ml.adapters;
 
 import java.io.File;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import org.integratedmodelling.kim.api.IParameters;
@@ -12,6 +14,7 @@ import org.integratedmodelling.klab.api.data.IResource.Builder;
 import org.integratedmodelling.klab.api.data.adapters.IResourceImporter;
 import org.integratedmodelling.klab.api.observations.IObservation;
 import org.integratedmodelling.klab.api.runtime.monitoring.IMonitor;
+import org.integratedmodelling.klab.utils.Triple;
 
 public class WekaImporter implements IResourceImporter {
 
@@ -29,13 +32,13 @@ public class WekaImporter implements IResourceImporter {
 	
 
     @Override
-    public Map<String, String> getExportCapabilities(IObservation observation) {
-        Map<String, String> ret = new HashMap<>();
+    public Collection<Triple<String, String, String>> getExportCapabilities(IObservation observation) {
+        List<Triple<String, String, String>> ret = new ArrayList<>();
         return ret;
     }
 
     @Override
-    public File exportObservation(File file, IObservation observation, ILocator locator, String format) {
+    public File exportObservation(File file, IObservation observation, ILocator locator, String format, IMonitor monitor) {
         // TODO Auto-generated method stub
         return null;
     }